requirements.txt
asgiref==3.8.1
attrs==25.1.0
base==0.0.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
whitenoise==6.9.0
pytest~=8.3.5